refactor(tau-coding): reduce coupling in session and TUI hotspots
Move terminal-command parsing, session-tree traversal, completion rendering, and completion layout into focused modules while preserving the session, app, and widgets compatibility facades for existing CLI, RPC, TUI, and test imports.
