+++ src/tau_coding/session_terminal.py
@@ -1,0 +1,53 @@
+"""Input-bar terminal command types and parsing.
+
+This module deliberately owns only the syntax and value objects.  Executing a
+command and persisting its result remain responsibilities of ``CodingSession``.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True)
+class TerminalCommandResult:
+    """Result of an input-bar terminal command."""
+
+    command: str
+    output: str
+    exit_code: int | None
+    ok: bool
+    added_to_context: bool
+
+
+@dataclass(frozen=True, slots=True)
+class TerminalCommandRequest:
+    """Parsed input-bar terminal command request."""
+
+    command: str
+    add_to_context: bool
+
+
+def terminal_command_context_message(command: str, output: str) -> str:
+    """Format terminal output for inclusion in agent context."""
+    return (
+        "Terminal command executed by the user.\n\n"
+        f"Command:\n```bash\n{command}\n```\n\n"
+        f"Output:\n```text\n{output}\n```"
+    )
+
+
+def parse_terminal_command(text: str) -> TerminalCommandRequest | None:
+    """Parse input-bar terminal command syntax."""
+    stripped = text.strip()
+    if stripped.startswith("!!"):
+        command = stripped[2:].strip()
+        if not command:
+            return None
+        return TerminalCommandRequest(command=command, add_to_context=False)
+    if stripped.startswith("!"):
+        command = stripped[1:].strip()
+        if not command:
+            return None
+        return TerminalCommandRequest(command=command, add_to_context=True)
+    return None
