+++ src/tau_coding/tui/completion_layout.py
@@ -1,0 +1,247 @@
+"""Prompt activity and completion-layout helpers for the Textual adapter."""
+
+from __future__ import annotations
+
+from io import StringIO
+
+from rich.console import Console
+from rich.text import Text
+from textual.widgets import Static
+
+from tau_agent.events import MessageEndEvent
+from tau_agent.messages import CustomMessage, UserMessage
+from tau_coding.events import CodingSessionEvent
+from tau_coding.tui.autocomplete import CompletionItem, CompletionState
+from tau_coding.tui.completion_widgets import render_completion_suggestions
+from tau_coding.tui.config import TAU_DARK_THEME, TuiTheme
+
+ACTIVITY_INDICATOR_HEIGHT = 3
+COMPLETION_MAX_VISIBLE_LINES = 16
+
+
+def activity_prompt_border_color(
+    theme: TuiTheme,
+    *,
+    frame: int,
+    running: bool,
+    shell_mode: bool,
+) -> str:
+    """Return the prompt border color for the current activity animation frame."""
+    del frame, running
+    if shell_mode:
+        return theme.role_styles["tool"].border
+    return theme.prompt_border
+
+
+def render_activity_indicator(
+    theme: TuiTheme,
+    *,
+    frame: int,
+    running: bool,
+    shell_mode: bool = False,
+) -> Text:
+    """Render the prompt prefix: a moving square while running, ``$`` in shell mode."""
+    if shell_mode and not running:
+        return Text("$", style=f"bold {theme.role_styles['tool'].border}")
+    if not running:
+        return Text("τ", style=f"bold {theme.accent}")
+
+    cycle_length = (ACTIVITY_INDICATOR_HEIGHT - 1) * 2
+    cycle_position = frame % cycle_length
+    active_row = (
+        cycle_position
+        if cycle_position < ACTIVITY_INDICATOR_HEIGHT
+        else cycle_length - cycle_position
+    )
+    direction = 1 if cycle_position < ACTIVITY_INDICATOR_HEIGHT else -1
+    trail_rows = {
+        active_row: theme.accent,
+        active_row - direction: blend_hex_colors(
+            theme.accent,
+            theme.screen_background,
+            fraction=0.35,
+        ),
+        active_row - (direction * 2): blend_hex_colors(
+            theme.accent,
+            theme.screen_background,
+            fraction=0.65,
+        ),
+    }
+
+    rendered = Text()
+    for row in range(ACTIVITY_INDICATOR_HEIGHT):
+        color = trail_rows.get(row)
+        if color is None:
+            rendered.append(" ")
+        else:
+            rendered.append("■", style=color)
+        if row < ACTIVITY_INDICATOR_HEIGHT - 1:
+            rendered.append("\n")
+    return rendered
+
+
+def is_terminal_command_prompt(text: str) -> bool:
+    """Return whether the prompt is currently in terminal-command mode."""
+    return terminal_command_prefix_span(text) is not None
+
+
+def should_optimistically_render_prompt(text: str) -> bool:
+    """Return whether submitted text can be safely shown before session expansion."""
+    stripped = text.strip()
+    return bool(stripped) and not stripped.startswith("/")
+
+
+def is_user_message_end_event(event: CodingSessionEvent) -> bool:
+    """Return whether an agent event closes a user-context message."""
+    return isinstance(event, MessageEndEvent) and isinstance(
+        event.message, (UserMessage, CustomMessage)
+    )
+
+
+def terminal_command_prefix_span(text: str) -> tuple[int, int] | None:
+    """Return the input span for a leading ! or !! terminal-command prefix."""
+    leading_whitespace = len(text) - len(text.lstrip())
+    stripped = text[leading_whitespace:]
+    if stripped.startswith("!!"):
+        return (leading_whitespace, leading_whitespace + 2)
+    if stripped.startswith("!"):
+        return (leading_whitespace, leading_whitespace + 1)
+    return None
+
+
+def blend_hex_colors(start: str, end: str, *, fraction: float) -> str:
+    """Blend two ``#rrggbb`` colors by ``fraction``."""
+    start_rgb = hex_to_rgb(start)
+    end_rgb = hex_to_rgb(end)
+    blended = tuple(
+        round(start_channel + (end_channel - start_channel) * fraction)
+        for start_channel, end_channel in zip(start_rgb, end_rgb, strict=True)
+    )
+    return f"#{blended[0]:02x}{blended[1]:02x}{blended[2]:02x}"
+
+
+def hex_to_rgb(color: str) -> tuple[int, int, int]:
+    """Parse a six-digit RGB hex color."""
+    value = color.removeprefix("#")
+    if len(value) != 6:
+        raise ValueError(f"Expected #rrggbb color, got {color!r}")
+    return (int(value[0:2], 16), int(value[2:4], 16), int(value[4:6], 16))
+
+
+def completion_visible_line_limit(suggestions: Static) -> int:
+    """Return the number of completion render lines that fit in the widget body."""
+    if suggestions.size.height > 0:
+        return max(min(COMPLETION_MAX_VISIBLE_LINES, suggestions.size.height), 1)
+    return COMPLETION_MAX_VISIBLE_LINES
+
+
+def visible_completion_state(
+    state: CompletionState,
+    *,
+    max_lines: int,
+    width: int | None = None,
+) -> CompletionState:
+    """Return a completion-state window with the selected item visible."""
+    if not state.items or max_lines <= 0:
+        return CompletionState()
+
+    selected_line_limit = max(max_lines - 1, 1)
+    start = 0
+    while start < state.selected_index:
+        candidate = CompletionState(
+            items=state.items[start:],
+            selected_index=state.selected_index - start,
+        )
+        if completion_selected_render_line(candidate, width=width) < selected_line_limit:
+            break
+        start += 1
+
+    end = len(state.items)
+    while end > state.selected_index + 1:
+        candidate = CompletionState(
+            items=state.items[start:end],
+            selected_index=state.selected_index - start,
+        )
+        if completion_render_line_count(candidate, width=width) <= max_lines:
+            break
+        end -= 1
+
+    while start < state.selected_index:
+        candidate = CompletionState(
+            items=state.items[start:end],
+            selected_index=state.selected_index - start,
+        )
+        if completion_render_line_count(candidate, width=width) <= max_lines:
+            break
+        start += 1
+
+    return CompletionState(
+        items=state.items[start:end],
+        selected_index=state.selected_index - start,
+    )
+
+
+def completion_selected_render_line(state: CompletionState, *, width: int | None = None) -> int:
+    """Return the rendered line number for the selected completion item."""
+    line = 0
+    has_rendered_text = False
+    previous_category: str | None = None
+    for index, item in enumerate(state.items):
+        if item.category != previous_category:
+            if has_rendered_text:
+                line += 1
+            if item.category:
+                line += 1
+                has_rendered_text = True
+            previous_category = item.category
+        elif has_rendered_text:
+            line += 1
+        if index == state.selected_index:
+            return line
+        line += completion_item_extra_wrapped_lines(item, width=width)
+        has_rendered_text = True
+    return line
+
+
+def completion_render_line_count(state: CompletionState, *, width: int | None = None) -> int:
+    """Return how many lines the completion state renders into."""
+    if not state.items:
+        return 0
+    line_count = 0
+    previous_category: str | None = None
+    for index, item in enumerate(state.items):
+        if item.category != previous_category:
+            if index:
+                line_count += 1
+            if item.category:
+                line_count += 1
+            previous_category = item.category
+        line_count += 1 + completion_item_extra_wrapped_lines(item, width=width)
+    return line_count
+
+
+def completion_item_extra_wrapped_lines(
+    item: CompletionItem,
+    *,
+    width: int | None,
+) -> int:
+    """Return extra rendered lines used when a completion description wraps."""
+    if width is None or width <= 0 or not item.description:
+        return 0
+    output = StringIO()
+    console = Console(
+        file=output,
+        width=width,
+        force_terminal=False,
+        color_system=None,
+        legacy_windows=False,
+    )
+    console.print(
+        render_completion_suggestions(
+            CompletionState(items=(item,), selected_index=0),
+            theme=TAU_DARK_THEME,
+        ),
+        end="",
+    )
+    line_count = len(output.getvalue().splitlines())
+    return max(line_count - 1, 0)
